refactor: 업로드 페이지와 공통인 부분 분리하기
